computer crashed
git-svn-id: https://svn.neumont.edu:8443/svn/csc190blake_rollins@65 e569a64c-822d-f44b-be61-ecb32803aff7

+++ GameSkeleton/GameSolution/Game/ExplosionEffect.h
@@ -7,7 +7,7 @@
 class ExplosionEffect : public ParticleEffect
 {
 public:
-	ExplosionEffect(Vector2D inOrigin, int inNumParticles, RGB inBaseColor, int inLifetime);
+	ExplosionEffect(Vector2D inOrigin, int inNumParticles, RGB inBaseColor, float inLifetime);
 	~ExplosionEffect(void);
 
 	void draw(Core::Graphics& g);
